finished shop

+++ src/components/Bin/index.js
@@ -5,7 +5,9 @@ import 'bootstrap/dist/css/bootstrap.css'
 
 export default class Bin extends Component {
     render() {
-        
+        var style = {
+            fontSize: 35
+          };
         var arr = [];
         var prod = [
             {
@@ -24,7 +26,7 @@ export default class Bin extends Component {
         console.log('key massiv: ',arr)
         console.log('product massiv: ',prod)
         console.log('totalcost: ',totalcost)
-        const listSales = prod.map((product) =>
+        const listSales = prod.map((product) => 
             <section key={product.id} className="product_section"> 
                             <li className="product_li title"> 
                                 <h2>
@@ -41,6 +43,7 @@ export default class Bin extends Component {
                                 </h4>
                             </li>
             </section>);
+            if(localStorage.length > 0){ 
             return (
                      <div className="container">
                         <div className="jumbotron">
@@ -54,15 +57,44 @@ export default class Bin extends Component {
                                 <ul>{listSales}</ul>
                             </div>
                             <div className="totalcost">
-                            <button className="btn" onClick={this.DeteleAllOrders}>Delete all orders</button>
+                            <Link to='/bin'><button className="btn" onClick={this.DeteleAllOrders}>Delete all orders</button></Link>
                             <br/>
                                 <span>Общая стоимость:</span> <span className="cost">{totalcost}</span> грн.
                             </div>
                         </div> 
                     </div>
-            )
+            )}
+            else {
+                return(
+                    <div className="container">
+                    <div className="jumbotron">
+                        <div className="title-order display-4">
+                            Your orders
+                            <br/>
+                            <hr/>
+                            <Link to='/'><img className="home-logo" alt="home" src="/img/home.png" /></Link>
+                            <br/>
+                                <br/>
+                            <br/>
+                            <div className="display-3">
+                                Your shopping cart is empty.
+                            </div>
+                            <br/>
+                            <div clasName="purpose" style={style}>
+                                You can buy whatever you want on the main page. 
+                            </div>
+                        </div>
+                    </div> 
+                </div>
+                )
+            }
         }
-        DeteleAllOrders = function() {
-            localStorage.clear()
+        DeteleAllOrders = () => {
+            if(localStorage.length > 0) {
+                localStorage.clear()
+            }
+            else {
+                alert('You have no orders.')
+            }
         }
     }
